move BottomSheet to shared directory
