+++ src/frontend/screens/ObservationEdit/index.tsx
@@ -6,7 +6,7 @@ import {usePersistedDraftObservation} from '../../hooks/persistedState/usePersis
 import {View, ScrollView, StyleSheet} from 'react-native';
 import {LocationView} from './LocationView';
 import {DescriptionField} from './DescriptionField';
-import {BottomSheet} from './BottomSheet';
+import {BottomSheet} from '../../sharedComponents/BottomSheet/BottomSheet';
 import {ThumbnailScrollView} from '../../sharedComponents/ThumbnailScrollView';
 import {PresetView} from './PresetView';
 import {useBottomSheetModal} from '../../sharedComponents/BottomSheetModal';
